Don't try to generate a config file when we don't have all the information
We don't want to store the database password for the sreview database in
cleartext in an open git repository, even if that database is firewalled
away. Therefore, people are expected to run ansible with '-e
sreviewdbpw=<the password>'. However, that breaks some important use
cases.

Avoid the problem by not generating the configuration file when we don't
have that password